pyproject.toml
[tool.poetry]
name = "kadastr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "kadastr"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
Jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
requests = "^2.32.3"
SQLAlchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
psycopg = {extras = ["binary"], version = "^3.2.1"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
